amn markazy audio looping, loader fix

+++ public/tsquare/js/nezal/display.js
@@ -11,10 +11,11 @@ var Display = Class.create({
 		Object.extend(this,properties)
     Object.extend(owner,this)
     owner.initDisplay()
-    
-    owner.scene.observe('end', function(){
-      owner.destroyAudio()
-    })
+    if (owner.scene) {
+      owner.scene.observe('end', function(){
+        owner.destroyAudio()
+      })
+    }
 	},
 	
   initDisplay : function(){
